chore: rearrange plots in analysis module
