+++ etna/analysis/__init__.py
@@ -1,42 +1,42 @@
-from etna.analysis.change_points_trend import find_change_points
-from etna.analysis.eda_utils import SeasonalPlotAggregation
-from etna.analysis.eda_utils import SeasonalPlotAlignment
-from etna.analysis.eda_utils import SeasonalPlotCycle
-from etna.analysis.eda_utils import acf_plot
-from etna.analysis.eda_utils import cross_corr_plot
-from etna.analysis.eda_utils import distribution_plot
-from etna.analysis.eda_utils import prediction_actual_scatter_plot
-from etna.analysis.eda_utils import qq_plot
-from etna.analysis.eda_utils import seasonal_plot
-from etna.analysis.eda_utils import stl_plot
+from etna.analysis.decomposition import SeasonalPlotAggregation
+from etna.analysis.decomposition import SeasonalPlotAlignment
+from etna.analysis.decomposition import SeasonalPlotCycle
+from etna.analysis.decomposition import find_change_points
+from etna.analysis.decomposition import plot_change_points_interactive
+from etna.analysis.decomposition import plot_time_series_with_change_points
+from etna.analysis.decomposition import plot_trend
+from etna.analysis.decomposition import seasonal_plot
+from etna.analysis.decomposition import stl_plot
+from etna.analysis.eda import acf_plot
+from etna.analysis.eda import cross_corr_plot
+from etna.analysis.eda import distribution_plot
+from etna.analysis.eda import get_correlation_matrix
+from etna.analysis.eda import plot_clusters
+from etna.analysis.eda import plot_correlation_matrix
+from etna.analysis.eda import plot_holidays
+from etna.analysis.eda import plot_periodogram
+from etna.analysis.feature_relevance import plot_feature_relevance
 from etna.analysis.feature_relevance.relevance import ModelRelevanceTable
 from etna.analysis.feature_relevance.relevance import RelevanceTable
 from etna.analysis.feature_relevance.relevance import StatisticsRelevanceTable
 from etna.analysis.feature_relevance.relevance_table import get_model_relevance_table
 from etna.analysis.feature_relevance.relevance_table import get_statistics_relevance_table
 from etna.analysis.feature_selection.mrmr_selection import AggregationMode
+from etna.analysis.forecast import get_residuals
+from etna.analysis.forecast import metric_per_segment_distribution_plot
+from etna.analysis.forecast import plot_backtest
+from etna.analysis.forecast import plot_backtest_interactive
+from etna.analysis.forecast import plot_forecast
+from etna.analysis.forecast import plot_forecast_decomposition
+from etna.analysis.forecast import plot_metric_per_segment
+from etna.analysis.forecast import plot_residuals
+from etna.analysis.forecast import prediction_actual_scatter_plot
+from etna.analysis.forecast import qq_plot
+from etna.analysis.outliers import plot_anomalies
+from etna.analysis.outliers import plot_anomalies_interactive
+from etna.analysis.outliers import plot_imputation
 from etna.analysis.outliers.density_outliers import absolute_difference_distance
 from etna.analysis.outliers.density_outliers import get_anomalies_density
 from etna.analysis.outliers.hist_outliers import get_anomalies_hist
 from etna.analysis.outliers.median_outliers import get_anomalies_median
 from etna.analysis.outliers.prediction_interval_outliers import get_anomalies_prediction_interval
-from etna.analysis.plotters import get_correlation_matrix
-from etna.analysis.plotters import get_residuals
-from etna.analysis.plotters import metric_per_segment_distribution_plot
-from etna.analysis.plotters import plot_anomalies
-from etna.analysis.plotters import plot_anomalies_interactive
-from etna.analysis.plotters import plot_backtest
-from etna.analysis.plotters import plot_backtest_interactive
-from etna.analysis.plotters import plot_change_points_interactive
-from etna.analysis.plotters import plot_clusters
-from etna.analysis.plotters import plot_correlation_matrix
-from etna.analysis.plotters import plot_feature_relevance
-from etna.analysis.plotters import plot_forecast
-from etna.analysis.plotters import plot_forecast_decomposition
-from etna.analysis.plotters import plot_holidays
-from etna.analysis.plotters import plot_imputation
-from etna.analysis.plotters import plot_metric_per_segment
-from etna.analysis.plotters import plot_periodogram
-from etna.analysis.plotters import plot_residuals
-from etna.analysis.plotters import plot_time_series_with_change_points
-from etna.analysis.plotters import plot_trend
