--- etna/analysis/change_points_trend/__init__.py
@@ -1,1 +1,0 @@
-from etna.analysis.change_points_trend.search import find_change_points
