+++ etna/analysis/decomposition/__init__.py
@@ -1,0 +1,9 @@
+from etna.analysis.decomposition.plots import plot_change_points_interactive
+from etna.analysis.decomposition.plots import plot_time_series_with_change_points
+from etna.analysis.decomposition.plots import plot_trend
+from etna.analysis.decomposition.plots import stl_plot
+from etna.analysis.decomposition.search import find_change_points
+from etna.analysis.decomposition.seasonal_plot import SeasonalPlotAggregation
+from etna.analysis.decomposition.seasonal_plot import SeasonalPlotAlignment
+from etna.analysis.decomposition.seasonal_plot import SeasonalPlotCycle
+from etna.analysis.decomposition.seasonal_plot import seasonal_plot
