+++ etna/analysis/decomposition/search.py
