+++ etna/analysis/eda/__init__.py
@@ -1,0 +1,8 @@
+from etna.analysis.eda.plots import acf_plot
+from etna.analysis.eda.plots import cross_corr_plot
+from etna.analysis.eda.plots import distribution_plot
+from etna.analysis.eda.plots import plot_clusters
+from etna.analysis.eda.plots import plot_correlation_matrix
+from etna.analysis.eda.plots import plot_holidays
+from etna.analysis.eda.plots import plot_periodogram
+from etna.analysis.eda.utils import get_correlation_matrix
