+++ etna/analysis/forecast/__init__.py
@@ -1,0 +1,10 @@
+from etna.analysis.forecast.plots import metric_per_segment_distribution_plot
+from etna.analysis.forecast.plots import plot_backtest
+from etna.analysis.forecast.plots import plot_backtest_interactive
+from etna.analysis.forecast.plots import plot_forecast
+from etna.analysis.forecast.plots import plot_forecast_decomposition
+from etna.analysis.forecast.plots import plot_metric_per_segment
+from etna.analysis.forecast.plots import plot_residuals
+from etna.analysis.forecast.plots import prediction_actual_scatter_plot
+from etna.analysis.forecast.plots import qq_plot
+from etna.analysis.forecast.utils import get_residuals
