+++ tests/test_analysis/conftest.py
@@ -1,0 +1,8 @@
+import matplotlib.pyplot as plt
+import pytest
+
+
+@pytest.fixture(autouse=True)
+def close_plots():
+    yield
+    plt.close()
