+++ tests/test_analysis/test_forecast/test_plots.py
@@ -1,0 +1,96 @@
+import pandas as pd
+import pytest
+
+from etna.analysis import plot_residuals
+from etna.analysis.forecast.plots import _validate_intersecting_segments
+from etna.metrics import MAE
+from etna.models import LinearPerSegmentModel
+from etna.pipeline import Pipeline
+from etna.transforms import LagTransform
+
+
+def test_plot_residuals_fails_unkown_feature(example_tsdf):
+    """Test that plot_residuals fails if meet unknown feature."""
+    pipeline = Pipeline(
+        model=LinearPerSegmentModel(), transforms=[LagTransform(in_column="target", lags=[5, 6, 7])], horizon=5
+    )
+    metrics, forecast_df, info = pipeline.backtest(ts=example_tsdf, metrics=[MAE()], n_folds=3)
+    with pytest.raises(ValueError, match="Given feature isn't present in the dataset"):
+        plot_residuals(forecast_df=forecast_df, ts=example_tsdf, feature="unkown_feature")
+
+
+@pytest.mark.parametrize(
+    "fold_numbers",
+    [
+        pd.Series([0, 0, 1, 1, 2, 2], index=pd.date_range("2020-01-01", periods=6, freq="D")),
+        pd.Series([0, 0, 1, 1, 2, 2], index=pd.date_range("2020-01-01", periods=6, freq="2D")),
+        pd.Series([2, 2, 0, 0, 1, 1], index=pd.date_range("2020-01-01", periods=6, freq="D")),
+        pd.Series(
+            [0, 0, 1, 1],
+            index=[
+                pd.Timestamp("2020-01-01"),
+                pd.Timestamp("2020-01-02"),
+                pd.Timestamp("2020-01-05"),
+                pd.Timestamp("2020-01-06"),
+            ],
+        ),
+    ],
+)
+def test_validate_intersecting_segments_ok(fold_numbers):
+    _validate_intersecting_segments(fold_numbers)
+
+
+@pytest.mark.parametrize(
+    "fold_numbers",
+    [
+        pd.Series(
+            [0, 0, 1, 1],
+            index=[
+                pd.Timestamp("2020-01-01"),
+                pd.Timestamp("2020-01-02"),
+                pd.Timestamp("2020-01-01"),
+                pd.Timestamp("2020-01-02"),
+            ],
+        ),
+        pd.Series([0, 0, 1, 1, 0, 0], index=pd.date_range("2020-01-01", periods=6, freq="D")),
+        pd.Series(
+            [0, 0, 1, 1],
+            index=[
+                pd.Timestamp("2020-01-01"),
+                pd.Timestamp("2020-01-02"),
+                pd.Timestamp("2020-01-02"),
+                pd.Timestamp("2020-01-03"),
+            ],
+        ),
+        pd.Series(
+            [1, 1, 0, 0],
+            index=[
+                pd.Timestamp("2020-01-01"),
+                pd.Timestamp("2020-01-02"),
+                pd.Timestamp("2020-01-02"),
+                pd.Timestamp("2020-01-03"),
+            ],
+        ),
+        pd.Series(
+            [0, 0, 1, 1],
+            index=[
+                pd.Timestamp("2020-01-01"),
+                pd.Timestamp("2020-01-05"),
+                pd.Timestamp("2020-01-03"),
+                pd.Timestamp("2020-01-08"),
+            ],
+        ),
+        pd.Series(
+            [1, 1, 0, 0],
+            index=[
+                pd.Timestamp("2020-01-01"),
+                pd.Timestamp("2020-01-05"),
+                pd.Timestamp("2020-01-03"),
+                pd.Timestamp("2020-01-08"),
+            ],
+        ),
+    ],
+)
+def test_validate_intersecting_segments_fail(fold_numbers):
+    with pytest.raises(ValueError):
+        _validate_intersecting_segments(fold_numbers)
